requirements.txt
streamlit
python-dotenv
openai
PyMuPDF
faiss-cpu